+++ core/src/main/java/react4j/Component.java
@@ -116,23 +116,6 @@ public abstract class Component
   protected abstract ReactNode render();
 
   /**
-   * The componentDidCatch() method works like a JavaScript catch {} block, but for components.
-   * Only class components can be error boundaries. In practice, most of the time you’ll want to
-   * declare an error boundary component once and use it throughout your application.
-   *
-   * <p>Note that error boundaries only catch errors in the components below them in the tree. An
-   * error boundary can’t catch an error within itself. If an error boundary fails trying to render
-   * the error message, the error will propagate to the closest error boundary above it. This, too,
-   * is similar to how catch {} block works in JavaScript.</p>
-   *
-   * @param error the error that has been thrown.
-   * @param info  information about component stack during thrown error.
-   */
-  protected void componentDidCatch( @Nonnull final JsError error, @Nonnull final ReactErrorInfo info )
-  {
-  }
-
-  /**
    * Store debug data on the state object of the native component.
    * This is only done if {@link React#shouldStoreDebugDataAsState()} returns true and is primarily
    * done to make it easy to debug the component from within React DevTools.