+++ processor/src/main/java/react4j/processor/Constants.java
@@ -3,10 +3,12 @@ package react4j.processor;
 final class Constants
 {
   static final String COMPONENT_CLASSNAME = "react4j.Component";
+  static final String ERROR_INFO_CLASSNAME = "react4j.ReactErrorInfo";
   static final String PROP_ANNOTATION_CLASSNAME = "react4j.annotations.Prop";
   static final String PROP_REF_ANNOTATION_CLASSNAME = "react4j.annotations.PropRef";
   static final String PROP_DEFAULT_ANNOTATION_CLASSNAME = "react4j.annotations.PropDefault";
   static final String PROP_VALIDATE_ANNOTATION_CLASSNAME = "react4j.annotations.PropValidate";
+  static final String ON_ERROR_ANNOTATION_CLASSNAME = "react4j.annotations.OnError";
   static final String ON_PROP_CHANGE_ANNOTATION_CLASSNAME = "react4j.annotations.OnPropChange";
   static final String REACT_COMPONENT_ANNOTATION_CLASSNAME = "react4j.annotations.ReactComponent";
   static final String PRE_UNMOUNT_ANNOTATION_CLASSNAME = "react4j.annotations.PreUnmount";
@@ -21,6 +23,7 @@ final class Constants
   static final String NONNULL_ANNOTATION_CLASSNAME = "javax.annotation.Nonnull";
   static final String NULLABLE_ANNOTATION_CLASSNAME = "javax.annotation.Nullable";
   static final String DEPRECATED_ANNOTATION_CLASSNAME = "java.lang.Deprecated";
+  static final String JS_ERROR_CLASSNAME = "elemental2.core.JsError";
 
   private Constants()
   {