+++ processor/src/test/resources/bad_input/com/example/on_error/AbstractOnErrorComponent.java
@@ -1,0 +1,23 @@
+package com.example.on_error;
+
+import elemental2.core.JsError;
+import javax.annotation.Nonnull;
+import react4j.Component;
+import react4j.ReactErrorInfo;
+import react4j.ReactNode;
+import react4j.annotations.OnError;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class AbstractOnErrorComponent
+  extends Component
+{
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @OnError
+  abstract void onError( @Nonnull JsError error, @Nonnull ReactErrorInfo info );
+}