+++ processor/src/test/resources/bad_input/com/example/on_error/ThrowsOnErrorComponent.java
@@ -1,0 +1,27 @@
+package com.example.on_error;
+
+import elemental2.core.JsError;
+import java.io.IOException;
+import javax.annotation.Nonnull;
+import react4j.Component;
+import react4j.ReactErrorInfo;
+import react4j.ReactNode;
+import react4j.annotations.OnError;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class ThrowsOnErrorComponent
+  extends Component
+{
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @OnError
+  void onError( @Nonnull final JsError error, @Nonnull final ReactErrorInfo info )
+    throws IOException
+  {
+  }
+}