+++ processor/src/test/resources/input/com/example/on_error/InfoOnlyOnErrorComponent.java
@@ -1,0 +1,24 @@
+package com.example.on_error;
+
+import javax.annotation.Nonnull;
+import react4j.Component;
+import react4j.ReactErrorInfo;
+import react4j.ReactNode;
+import react4j.annotations.OnError;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class InfoOnlyOnErrorComponent
+  extends Component
+{
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @OnError
+  void onError( @Nonnull final ReactErrorInfo info )
+  {
+  }
+}